Shuffle tests around and split utils into smaller modules
