+++ crates/tako/src/scheduler/metrics.rs
@@ -49,7 +49,7 @@ fn crawl<F1: Fn(&Task) -> &Set<TaskRef>, F2: Fn(&Task) -> &Vec<TaskRef>>(
 mod tests {
     use crate::scheduler::metrics::compute_b_level_metric;
     use crate::server::core::Core;
-    use crate::tests::utils::submit_example_2;
+    use crate::tests::utils::workflows::submit_example_2;
 
     #[test]
     fn b_level_simple_graph() {
